# src/internal/docs/swagger.yaml
basePath: /api
definitions:
  db.Role:
    enum:
    - member
    - admin
    type: string
    x-enum-varnames:
    - RoleMember
    - RoleAdmin
  user.ProfileDto:
    properties:
      createdAt:
        type: string
      email:
        example: test@test.com
        type: string
      firstName:
        example: John
        type: string
      id:
        format: int64
        type: integer
      lastName:
        example: Wick
        type: string
      userRole:
        allOf:
        - $ref: '#/definitions/db.Role'
        example: member
    type: object
  user.SignInDto:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: test123
        type: string
    type: object
  user.SignUpDto:
    properties:
      email:
        example: test@test.com
        type: string
      firstName:
        example: John
        type: string
      lastName:
        example: Wick
        type: string
      password:
        example: test123
        type: string
    type: object
  user.TokenDto:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
  description: Swagger for Shop Api
  title: Shop Api
  version: "1.0"
paths:
  /users/get:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.ProfileDto'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get my profile
      tags:
      - users
  /users/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ProfileDto'
      security:
      - ApiKeyAuth: []
      summary: Get my profile
      tags:
      - users
  /users/signin:
    post:
      parameters:
      - description: User details for login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignInDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.TokenDto'
      summary: Sign in
      tags:
      - users
  /users/signup:
    post:
      parameters:
      - description: User details for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.TokenDto'
      summary: Sign Up
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"